+++ scripts/stubs.sh
@@ -1,0 +1,3 @@
+#! /bin/bash
+stubgen src/dependency -o stubs/dependency --include-docstrings
+stubgen src/library -o stubs/library --include-docstrings