+++ src/example/plugin/base/imports.py
@@ -4,4 +4,4 @@ import example.plugin.base.string.providers.fake
 
 # Plugin import
 from example.plugin.base import BasePlugin
-__all__ = [BasePlugin]
+__all__ = ["BasePlugin"]