+++ src/example/plugin/reporter/imports.py
@@ -4,4 +4,4 @@ import example.plugin.reporter.facade.providers.facadeA
 
 # Plugin import
 from example.plugin.reporter import ReporterPlugin
-__all__ = [ReporterPlugin]
+__all__ = ["ReporterPlugin"]